+++ noodles-gff/src/feature/record_buf/convert.rs
@@ -13,7 +13,7 @@ impl RecordBuf {
 
         builder = builder
             .set_reference_sequence_name(record.reference_sequence_name().into_owned())
-            .set_source(record.source())
+            .set_source(record.source().into_owned())
             .set_type(record.ty())
             .set_start(record.feature_start()?)
             .set_end(record.feature_end()?);