+++ noodles-gtf/src/io/writer/line/record.rs
@@ -25,7 +25,7 @@ where
     write_reference_sequence_name(writer, record.reference_sequence_name().as_ref())?;
 
     write_separator(writer)?;
-    write_source(writer, record.source())?;
+    write_source(writer, record.source().as_ref())?;
 
     write_separator(writer)?;
     write_type(writer, record.ty())?;