+++ noodles-gtf/src/record.rs
@@ -93,8 +93,8 @@ impl gff::feature::Record for Record<'_> {
         Cow::from(self.reference_sequence_name())
     }
 
-    fn source(&self) -> &BStr {
-        self.source()
+    fn source(&self) -> Cow<'_, BStr> {
+        Cow::from(self.source())
     }
 
     fn ty(&self) -> &BStr {